#10076: Merge acosh_bw to TTNN
