+++ docs/source/ttnn/ttnn/api.rst
@@ -178,6 +178,7 @@ Pointwise Unary
    ttnn/fill_bw
    ttnn/hardsigmoid_bw
    ttnn/cos_bw
+   ttnn/acosh_bw
    ttnn/sub_bw
    ttnn/frac_bw
    ttnn/trunc_bw
