+++ docs/source/ttnn/ttnn/ttnn/acosh_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.acosh_bw:
+
+ttnn.acosh_bw
+###############
+
+.. autofunction:: ttnn.acosh_bw
