+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -200,11 +200,6 @@ std::vector<Tensor> cosh_bw(
     const Tensor& input,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<Tensor> acosh_bw(
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
-
 std::vector<Tensor> acos_bw(
     const Tensor& grad,
     const Tensor& input,
